fixed models from long to int, still producing multiple of same tables

+++ DAL/Models/Exercise.cs
@@ -4,9 +4,9 @@ namespace workout_app.DAL.Models
 {
     public class Exercise
     {
-        public long ExerciseId { get; set; }
+        public int ExerciseId { get; set; }
         public string ExerciseName { get; set; }
-        public long MuscleGroupId { get; set; }
+        public int MuscleGroupId { get; set; }
 
         public virtual MuscleGroup MuscleGroups { get; set; }
         public virtual ICollection<Workout> Workouts { get; private set; }
