+++ DAL/Models/MuscleGroup.cs
@@ -4,7 +4,7 @@ namespace workout_app.DAL.Models
 {
     public class MuscleGroup
     {
-        public long MuscleGroupId { get; set; }
+        public int MuscleGroupId { get; set; }
         public string MuscleGroupName { get; set; }  
 
          public virtual ICollection<Exercise> Exercises { get; private set;}
