+++ ViewModels/ExcersiseViewModel.cs
@@ -2,9 +2,9 @@ namespace workout_app.ViewModels
 {
     public class ExerciseViewModel
     {
-        public long ExerciseId { get; set; }
+        public int ExerciseId { get; set; }
         public string ExerciseName { get; set; }
-        public long MuscleGroupId { get; set; }
+        public int MuscleGroupId { get; set; }
 
     }
 }
