+++ ViewModels/MuscleGroupViewModel.cs
@@ -2,8 +2,7 @@ namespace workout_app.ViewModels
 {
     public class MuscleGroupViewModel
     {        
-        public long MuscleGroupId { get; set; }
+        public int MuscleGroupId { get; set; }
         public string MuscleGroupName { get; set; }
-
     }
 }
